requirements.txt
Flask==2.2.5
   torch==2.7.1
   torchvision==0.22.1
   Pillow==10.4.0
   requests==2.32.3
   flask-cors==4.0.1
   gunicorn==22.0.0